Moved tests, made javascript files consistent and converted them to plugins
