+++ frontend/package.json
@@ -131,7 +131,7 @@
     ],
     "testMatch": [
       "**/tests/unit/**/*.spec.(js|jsx|ts|tsx)",
-      "**/__tests__/*.(js|jsx|ts|tsx)"
+      "**/__tests__/*.spec.(js|jsx|ts|tsx)"
     ],
     "testURL": "http://localhost/",
     "coverageDirectory": "data/coverage",
