+++ frontend/src/plugins/index.js
@@ -1,0 +1,4 @@
+export { default as vuetify } from './vuetify'
+export { default as store } from './store'
+export { default as auth } from './auth'
+export { default as filterLoading } from './filterLoading'
