+++ frontend/src/plugins/store/__tests__/.eslintrc.js
