+++ frontend/src/plugins/store/__tests__/resources/circular-reference.json
