+++ frontend/src/plugins/store/__tests__/resources/collection-complete.json
