+++ frontend/src/plugins/store/__tests__/resources/collection-page1.json
