+++ frontend/src/plugins/store/__tests__/resources/embedded-collection.json
