+++ frontend/src/plugins/store/__tests__/resources/embedded-single-entity.json
