+++ frontend/src/plugins/store/__tests__/resources/linked-collection.json
