+++ frontend/src/plugins/store/__tests__/resources/linked-single-entity.json
