+++ frontend/src/plugins/store/__tests__/resources/multiple-references-to-user.json
