+++ frontend/src/plugins/store/__tests__/resources/reference-to-single-entity.json
