+++ frontend/src/plugins/store/__tests__/resources/templated-link.json
