+++ frontend/src/plugins/store/normalizeUri.js
