+++ frontend/src/plugins/vuetify.js
@@ -1,7 +1,11 @@
+// You still need to register Vuetify itself
+// src/plugins/vuetify.js
+
 import Vue from 'vue'
 import Vuetify from 'vuetify/lib'
 
 Vue.use(Vuetify)
 
-export default new Vuetify({
-})
+const opts = {}
+
+export default new Vuetify(opts)
