+++ frontend/src/router.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import Router from 'vue-router'
 import slugify from 'slugify'
-import { refreshLoginStatus } from '@/auth/auth'
+import { refreshLoginStatus } from '@/plugins/auth'
 
 Vue.use(Router)
